Refaktor. Flytt dto til ny mappe
