+++ src/test/kotlin/no/nav/tilleggsstonader/sak/util/VilkårGrunnlagUtil.kt
@@ -1,18 +1,18 @@
 package no.nav.tilleggsstonader.sak.util
 
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.GrunnlagAktivitet
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.GrunnlagBarn
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.GrunnlagHovedytelse
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.RegistergrunnlagBarn
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.SøknadsgrunnlagBarn
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.VilkårGrunnlagDto
+import no.nav.tilleggsstonader.sak.behandling.fakta.BehandlingFaktaDto
+import no.nav.tilleggsstonader.sak.behandling.fakta.GrunnlagAktivitet
+import no.nav.tilleggsstonader.sak.behandling.fakta.GrunnlagBarn
+import no.nav.tilleggsstonader.sak.behandling.fakta.GrunnlagHovedytelse
+import no.nav.tilleggsstonader.sak.behandling.fakta.RegistergrunnlagBarn
+import no.nav.tilleggsstonader.sak.behandling.fakta.SøknadsgrunnlagBarn
 import java.util.UUID
 
 object VilkårGrunnlagUtil {
     fun mockVilkårGrunnlagDto(
         barn: List<GrunnlagBarn> = emptyList(),
     ) =
-        VilkårGrunnlagDto(
+        BehandlingFaktaDto(
             hovedytelse = GrunnlagHovedytelse(
                 søknadsgrunnlag = null,
             ),
